Update README.md
This add-in is alternative to Excel graphics which facilitates use of GnuPlot directly in Excel sheets. GnuPlot scripts are stored in Excel along with the generated graphs and can be instantly edited and regenerated without leaving the Excel environment. Data can be read from Excel sheets by a cell range ("A1:C400") or using an Excel named range.